--- src/components/layout/introduce/index.tsx
@@ -1,40 +1,0 @@
-import { disableBody } from "@/utils/disableBody";
-import { MouseEvent, useCallback, useEffect, useState } from "react";
-import { BackgroundWrapper, Header, Present } from "..";
-
-const Introduce = () => {
-  const [open, setOpen] = useState(false);
-
-  const handleClick = useCallback(
-    (event: MouseEvent<HTMLDivElement>) => {
-      const target = event.target as HTMLDivElement;
-      const currentTarget = event.currentTarget as HTMLDivElement;
-
-      if (innerWidth > 730) return;
-
-      if (currentTarget.className.includes("burgerIcon")) {
-        setOpen(!open);
-      } else if (
-        !target.className.includes("menu") &&
-        !target.className.includes("burgerIcon") &&
-        open
-      ) {
-        setOpen(false);
-      }
-    },
-    [open]
-  );
-
-  useEffect(() => {
-    disableBody(open);
-  }, [open]);
-
-  return (
-    <BackgroundWrapper>
-      <Header open={open} handleClick={handleClick} />
-      <Present />
-    </BackgroundWrapper>
-  );
-};
-
-export default Introduce;
