+++ src/pages/mainPage/styles.module.scss
@@ -1,0 +1,9 @@
+.backgroundStyle {
+  background: url(noise_lines_transparent@2X),
+    radial-gradient(
+        234.05% 111.8% at 0% 0%,
+        rgb(91, 72, 58),
+        rgb(38, 36, 37) 100%
+      )
+      rgb(33, 31, 32);
+}
